category id field fixed

+++ src/main/java/com/core/entity/Category.java
@@ -10,7 +10,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Byte id;
+    private Integer id;
 
     @Column(name = "name")
     private String name;
